+++ netket/hilbert/qubit.py
@@ -38,7 +38,7 @@ class Qubit(HomogeneousHilbert):
             >>> print(hi.size)
             100
         """
-        super().__init__(StaticRange(0, 1, 2), N)
+        super().__init__(StaticRange(0, 1, 2, dtype=float), N)
 
     def states_to_local_indices(self, x):
         return x.astype(np.int32)
